renamed payout to betting
renamed payout to betting
