+++ app/models/behaviors/nonplayer/BlackjackNonPlayer.scala
@@ -2,7 +2,7 @@ package cards.models.behaviors.nonplayer
 
 import cards.models.behaviors.evaluation.BlackjackHandEvaluation
 import cards.models.behaviors.predicates.BlackjackPredicates
-import cards.models.behaviors.payout.BlackjackPayout
+import cards.models.behaviors.betting.BlackjackBetting
 import cards.models.classes.state.{ BlackjackPlayerState, BlackjackGameState }
 import cards.models.classes.{ Card, Deck }
 import cards.models.classes.actions.{ Action, BlackjackAction }
@@ -11,10 +11,10 @@ import cards.models.classes.actions.BlackjackAction._
 trait BlackjackNonPlayer { 
   type EVAL <: BlackjackHandEvaluation 
   type PREDICATES <: BlackjackPredicates
-  type PAYOUT <: BlackjackPayout
+  type BETTING <: BlackjackBetting
   val evaluation: EVAL
   val predicates: PREDICATES
-  val payout: PAYOUT
+  val betting: BETTING
 
   def next(gameState: BlackjackGameState): BlackjackGameState = {
     if (gameState.deck.length == 0) {
@@ -25,9 +25,10 @@ trait BlackjackNonPlayer {
       throw new IllegalStateException(
         s"Cannot get next because there are no players")
     }
-    if (payout.isTimeToSettle(gameState)) {
-      return payout.settleBets(gameState)
+    if (betting.isTimeToSettle(gameState)) {
+      return betting.settleBets(gameState)
     }
+    // TODO: isTimeToPlaceBets
 
     ???
   }
